feat: separate components and data from app page
