+++ src/components/FilteredEventList.tsx
@@ -1,0 +1,72 @@
+import { Notifications, Edit, Delete } from '@mui/icons-material';
+import { Box, Stack, Typography, IconButton } from '@mui/material';
+
+import { notificationOptions } from '../data';
+import { Event } from '../types';
+
+export const FilteredEventList = ({
+  filteredEvents,
+  notifiedEvents,
+  editEvent,
+  deleteEvent,
+}: {
+  filteredEvents: Event[];
+  notifiedEvents: string[];
+  editEvent: (event: Event) => void;
+  deleteEvent: (id: string) => Promise<void>;
+}) => {
+  return (
+    <>
+      {filteredEvents.map((event) => (
+        <Box key={event.id} sx={{ border: 1, borderRadius: 2, p: 3, width: '100%' }}>
+          <Stack direction="row" justifyContent="space-between">
+            <Stack>
+              <Stack direction="row" spacing={1} alignItems="center">
+                {notifiedEvents.includes(event.id) && <Notifications color="error" />}
+                <Typography
+                  fontWeight={notifiedEvents.includes(event.id) ? 'bold' : 'normal'}
+                  color={notifiedEvents.includes(event.id) ? 'error' : 'inherit'}
+                >
+                  {event.title}
+                </Typography>
+              </Stack>
+              <Typography>{event.date}</Typography>
+              <Typography>
+                {event.startTime} - {event.endTime}
+              </Typography>
+              <Typography>{event.description}</Typography>
+              <Typography>{event.location}</Typography>
+              <Typography>카테고리: {event.category}</Typography>
+              {event.repeat.type !== 'none' && (
+                <Typography>
+                  반복: {event.repeat.interval}
+                  {event.repeat.type === 'daily' && '일'}
+                  {event.repeat.type === 'weekly' && '주'}
+                  {event.repeat.type === 'monthly' && '월'}
+                  {event.repeat.type === 'yearly' && '년'}
+                  마다
+                  {event.repeat.endDate && ` (종료: ${event.repeat.endDate})`}
+                </Typography>
+              )}
+              <Typography>
+                알림:{' '}
+                {
+                  notificationOptions.find((option) => option.value === event.notificationTime)
+                    ?.label
+                }
+              </Typography>
+            </Stack>
+            <Stack>
+              <IconButton aria-label="Edit event" onClick={() => editEvent(event)}>
+                <Edit />
+              </IconButton>
+              <IconButton aria-label="Delete event" onClick={() => deleteEvent(event.id)}>
+                <Delete />
+              </IconButton>
+            </Stack>
+          </Stack>
+        </Box>
+      ))}
+    </>
+  );
+};
